Revert "Change highlight slide link to be direct download"
This reverts commit e655359b1e91a8bcf33261b6d002d12c0ac89018.

I was originally going to make all the links to slides like
highlights and presentation direct downloads for PowerPoint
files. I personally like these better because I don't have
to go to a secondary link to download the file.

However, not everyone has PowerPoint, so maybe getting a link
to the Google office tools would be helpful. Also, this gives
users a chance to ask for access if there is a permissions
error preventing downloads.